[Update] 연산 범위 증가

+++ Assets/Scripts/PlayerStats.cs
@@ -24,7 +24,7 @@ public class PlayerStats : MonoBehaviour
     [SerializeField]
     private Transform button;
     
-    public int gold;
+    public long gold;
 
     [SerializeField]
     private TextMeshProUGUI upgradeText;
@@ -189,8 +189,8 @@ public class PlayerStats : MonoBehaviour
         return stats;
     }
 
-    public void GetGold(float getGold) {
-        gold += (int)getGold;
+    public void GetGold(double getGold) {
+        gold += (long)getGold;
         UpdateGoldText();
     }
